requirements.txt
Django==1.11.29
django-bootstrap3==8.2.2
django-environ==0.4.3
django-pytest==0.2.0
future==0.16.0
gitdb==4.0.5
GitPython==3.1.11
jsonfield==2.0.1
nose==1.3.7
packaging==16.8
parsedatetime==2.3
psycopg2==2.8.5
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
python-dateutil==2.6.0
pytz==2020.4
six==1.10.0
smmap==3.0.4
sqlparse==0.2.3
svn==0.3.44
